fix: 결제취소 오류 수정, 파일명 수정, 목록gap 수정
