+++ app/(main)/cart/onetimeCart.tsx
@@ -1,4 +1,4 @@
-import SubscriptionItemList from "@/app/(main)/cart/_components/subscription-item-list";
+import OnetimeItemList from "@/app/(main)/cart/_components/onetimeItemList";
 import { deleteCartItem, deleteCartItems } from "@/app/(main)/cart/action/cart";
 import useCartStore from "@/zustand/useCartStore";
 import Button from "@/components/common/Button";
@@ -8,7 +8,7 @@ import Image from "next/image";
 import { useEffect, useMemo, useState } from "react";
 import { useRouter } from "next/navigation";
 
-export default function SubscriptionCart() {
+export default function OnetimeCart() {
   const router = useRouter();
 
   // 토큰 가져오기
@@ -20,7 +20,7 @@ export default function SubscriptionCart() {
     handleDeleteMultiple: deleteStoreItems,
     handleDeleteSuccess: deleteStoreItem,
     getCartTotal,
-    getSubscriptionItems,
+    getOnetimeItems,
     getSelectCartTotal,
     setCheckoutItems,
     isSoldOut,
@@ -30,37 +30,36 @@ export default function SubscriptionCart() {
   // 체크박스 선택된 상품 ID
   const [selectIds, setSelectIds] = useState<number[]>([]);
 
+  // 1회구매 상품 가져오기
+  const items = getOnetimeItems();
+
   // 컴포넌트 최초 로드 시 전체 선택 처리
   useEffect(() => {
-    const subItems = getSubscriptionItems();
-    if (subItems.length > 0) {
-      const allIds = subItems.map((item) => item._id);
+    if (items.length > 0) {
+      // 모든 상품의 ID를 추출하여 선택 상태로 설정
+      const allIds = items.map((item) => item._id);
       setSelectIds(allIds);
     }
-  }, [getSubscriptionItems().length]);
+  }, [items.length]);
 
-  // 정기구독 상품 가져오기
-  const items = getSubscriptionItems();
-
-  // 정기구독 총액 계산
-  const { productsPrice, shippingFees, totalPrice, availableCount, discount, selectCount } =
-    useMemo(() => {
-      if (selectIds.length === 0) {
-        return {
-          productsPrice: 0,
-          shippingFees: 0,
-          discount: 0,
-          totalPrice: 0,
-          selectCount: 0,
-          availableCount: getCartTotal("subscription").availableCount,
-        };
-      }
-      // 선택된 상품이 있을 때 계산
+  // 1회구매 총액 계산 (선택 || 전체)
+  const { productsPrice, shippingFees, totalPrice, selectCount, availableCount } = useMemo(() => {
+    if (selectIds.length === 0) {
       return {
-        ...getSelectCartTotal(selectIds, "subscription"),
-        availableCount: getCartTotal("subscription").availableCount,
+        productsPrice: 0,
+        shippingFees: 0,
+        totalPrice: 0,
+        selectCount: 0,
+        availableCount: getCartTotal("oneTime").availableCount,
       };
-    }, [cartData, selectIds, getSelectCartTotal, getCartTotal]);
+    }
+
+    // 선택된 상품이 있는 경우 계산
+    return {
+      ...getSelectCartTotal(selectIds, "oneTime"),
+      availableCount: getCartTotal("oneTime").availableCount,
+    };
+  }, [cartData, selectIds, getSelectCartTotal, getCartTotal]);
 
   // 한건 삭제 핸들러
   const handleDelete = async (cartId: number) => {
@@ -124,7 +123,7 @@ export default function SubscriptionCart() {
       if (result === null) {
         deleteStoreItems(selectIds);
 
-        setSelectIds([]);
+        setSelectIds([]); // 선택 목록 초기화
       } else {
         alert(result.message);
       }
@@ -137,16 +136,17 @@ export default function SubscriptionCart() {
 
   // 구매하기 버튼
   const handlePurchase = () => {
-    const subscriptionItems = getSubscriptionItems();
+    const onetimeItems = getOnetimeItems();
 
     let itemsToPurchase;
     if (selectIds.length > 0) {
       // 선택 상품(품절 제외)
-      itemsToPurchase = subscriptionItems.filter(
+      itemsToPurchase = onetimeItems.filter(
         (item) => selectIds.includes(item._id) && !isSoldOut(item._id)
       );
     } else {
-      itemsToPurchase = subscriptionItems.filter((item) => !isSoldOut(item._id));
+      // 전체 상품(품절 제외)
+      itemsToPurchase = onetimeItems.filter((item) => !isSoldOut(item._id));
     }
 
     if (itemsToPurchase.length === 0) {
@@ -172,6 +172,7 @@ export default function SubscriptionCart() {
                 onChange={handleSelectAll}
                 className="text-[#1A1A1C] text-[0.75rem] font-black"
               />
+
               <button
                 onClick={handleDeleteMultiple}
                 disabled={selectIds.length === 0 || isDeleting}
@@ -182,20 +183,22 @@ export default function SubscriptionCart() {
             </section>
 
             {/* 상품 목록 */}
-            {items.map((cart) => (
-              <SubscriptionItemList
-                key={cart._id}
-                cart={cart}
-                isSelect={selectIds.includes(cart._id)}
-                onSelect={() => handleSelect(cart._id)}
-                onDelete={handleDelete}
-              />
-            ))}
+            <div className="flex flex-col gap-3.5">
+              {items.map((cart) => (
+                <OnetimeItemList
+                  key={cart._id}
+                  cart={cart}
+                  isSelect={selectIds.includes(cart._id)}
+                  onSelect={() => handleSelect(cart._id)}
+                  onDelete={handleDelete}
+                />
+              ))}
+            </div>
           </>
         ) : (
-          <div className="border border-[#F9F9FB] rounded-[0.875rem] px-7 py-7 sm:px-7 sm:py-7 bg-white shadow-(--shadow-card)">
+          <div className="border border-[#F9F9FB] rounded-[0.875rem] px-7 py-7 xl:px-7 xl:py-20 bg-white shadow-(--shadow-card)">
             <p className="text-[0.75rem] text-[#1A1A1C] font-bold text-center">
-              구독 중인 상품이 없습니다.
+              장바구니에 담긴 상품이 없습니다.
             </p>
           </div>
         )}
@@ -218,27 +221,21 @@ export default function SubscriptionCart() {
                 {shippingFees.toLocaleString()}원
               </p>
             </div>
-            <div className="flex justify-between">
-              <p className="text-[0.75rem] text-text-secondary font-bold">정기구독 할인</p>
-              <p className="text-[0.75rem] text-(--color-accent-primary) font-black">
-                -{discount.toLocaleString()}원
-              </p>
-            </div>
 
             <div className="flex justify-between border-t border-border-primary py-7">
               <h2 className="text-[1rem] text-[#1A1A1C] font-black">총 결제 예정액</h2>
-              <p className="text-2xl text-[#FBA613] font-black">
-                {(totalPrice - discount).toLocaleString()}원
-              </p>
+              <p className="text-2xl text-[#FBA613] font-black">{totalPrice.toLocaleString()}원</p>
             </div>
 
             {/* 구매하기 버튼 */}
             <Button onClick={handlePurchase} disabled={selectIds.length === 0 || selectCount === 0}>
-              {selectIds.length === 0
-                ? "구매할 상품을 선택해주세요"
-                : selectCount > 0
-                ? `${selectCount}개 상품 구독하기`
-                : "구독 가능한 상품이 없습니다."}
+              {selectIds.length > 0
+                ? selectCount > 0
+                  ? `${selectCount}개 상품 구매하기`
+                  : "선택한 상품 중 구매 가능한 상품이 없습니다."
+                : availableCount > 0
+                ? `${availableCount}개 상품 구매하기`
+                : "구매 가능한 상품이 없습니다."}
             </Button>
 
             <div className="flex items-center justify-center gap-2">
@@ -250,15 +247,12 @@ export default function SubscriptionCart() {
           </section>
 
           {/* 정기 구독 혜택 */}
-          <section className="flex flex-col bg-[#FFF9F2] px-5 py-5 rounded-[0.875rem] gap-3 border border-[#FFF5E6]">
+          <section className="flex flex-col bg-[#FFF9F2] p-5 rounded-[0.875rem] gap-3 border border-[#FFF5E6]">
             <div className="flex items-center gap-2.5">
               <Image src="/images/cart/구독혜택.svg" alt="" width={28} height={28} />
               <h3 className="text-[0.75rem] text-[#1A1A1C] font-black">나만의 정기 구독 혜택</h3>
             </div>
             <ul className="flex flex-col gap-1.5 ml-4">
-              <li className="text-[0.75rem] text-text-tertiary font-bold list-disc ml-2 pl-1">
-                배송비 무료 혜택
-              </li>
               <li className="text-[0.75rem] text-text-tertiary font-bold list-disc ml-2 pl-1">
                 전 상품 10% 자동 상시 할인
               </li>
